fixed the error

+++ train_baseline.py
@@ -115,7 +115,7 @@ def train_baseline(
         logger=logger,
         log_every_n_steps=50,
         val_check_interval=1.0,
-        precision='16-mixed' if use_gpu else 32,  # Use mixed precision on GPU
+        precision=16 if use_gpu else 32,  # Use mixed precision on GPU
         deterministic=True,  # For reproducible results
         sync_batchnorm=True if (use_gpu and len(devices) > 1) else False  # Sync batch norm for multi-GPU
     )
